+++ beetsplug/ftintitle.py
@@ -49,13 +49,15 @@ def contains_feat(title):
 
 
 def update_metadata(item, feat_part):
-    """Choose how to add new artists to the title and write the new
+    """Choose how to add new artists to the title and set the new
     metadata. Also, print out messages about any changes that are made.
     """
     # In all cases, update the artist fields.
     ui.print_(u'artist: {0} -> {1}'.format(item.artist, item.albumartist))
     item.artist = item.albumartist
-    item.artist_sort, _ = split_on_feat(item.artist_sort)  # Strip featured.
+    if item.artist_sort:
+        # Just strip the featured artist from the sort name.
+        item.artist_sort, _ = split_on_feat(item.artist_sort)
 
     # Only update the title if it does not already contain a featured
     # artist.